UI Menu & Highlight
added a menu canvas that floats on the head of the pet dog, which can be hidden by clicking an icon toggle. And also set a button to make the dog follow or stay. Added a highlight affordance for several items, others are not suitable due to their multiple children.